refactor(chat): extract event classes from ChatStore to separate files
Move inner event classes (StoreUpdateEvent, CurrentChannelChangeEvent, ChannelsUpdateEvent, MessagesUpdateEvent, UsersUpdateEvent, UnreadUpdateEvent, LoadingMessagesEvent, ChatStateChange, ChatMessageReceive) from ChatStore into individual files within the events package. Update all imports accordingly to maintain functionality while improving code organization and separation of concerns.
