+++ src/mindustrytool/features/chat/global/events/ChannelsUpdateEvent.java
@@ -1,0 +1,4 @@
+package mindustrytool.features.chat.global.events;
+
+public class ChannelsUpdateEvent {
+}
