+++ src/mindustrytool/features/chat/global/events/ChatMessageReceive.java
@@ -1,7 +1,8 @@
-package mindustrytool.features.chat.global.dto;
+package mindustrytool.features.chat.global.events;
 
 import arc.struct.Seq;
 import lombok.RequiredArgsConstructor;
+import mindustrytool.features.chat.global.dto.ChatMessage;
 
 @RequiredArgsConstructor
 public class ChatMessageReceive {
