+++ src/mindustrytool/features/chat/global/events/CurrentChannelChangeEvent.java
@@ -1,0 +1,9 @@
+package mindustrytool.features.chat.global.events;
+
+public class CurrentChannelChangeEvent {
+    public final String channelId;
+
+    public CurrentChannelChangeEvent(String channelId) {
+        this.channelId = channelId;
+    }
+}
