+++ src/mindustrytool/features/chat/global/events/LoadingMessagesEvent.java
@@ -1,0 +1,9 @@
+package mindustrytool.features.chat.global.events;
+
+public class LoadingMessagesEvent {
+    public final boolean isLoading;
+
+    public LoadingMessagesEvent(boolean isLoading) {
+        this.isLoading = isLoading;
+    }
+}
