+++ src/mindustrytool/features/chat/global/events/StoreUpdateEvent.java
@@ -1,0 +1,4 @@
+package mindustrytool.features.chat.global.events;
+
+public class StoreUpdateEvent {
+}
