+++ src/mindustrytool/features/chat/global/events/UsersUpdateEvent.java
@@ -1,0 +1,9 @@
+package mindustrytool.features.chat.global.events;
+
+public class UsersUpdateEvent {
+    public final String channelId;
+
+    public UsersUpdateEvent(String channelId) {
+        this.channelId = channelId;
+    }
+}
